Phases 8-11: Code cleanup and refactoring
This commit implements phases 8-11 of the codebase maturation process:

**Phase 8: Remove Dead Code**
- Removed 518 lines of duplicate methods from AstroFileManager.py
- Methods were previously extracted to tab classes but left in main file
- Removed: on_keyword_changed, populate_current_values, replace_values,
  preview_organization, execute_organization, create_settings_tab,
  browse_repository, save_timezone_setting, apply_theme_setting

**Phase 9: Extract Utility Functions**
- Created new utils/ package with file_organizer.py module
- Moved generate_organized_path() function to utils module
- Improved code organization and reusability

**Phase 10: Fix Circular Import**
- Resolved circular import in MaintenanceTab
- Updated imports to use utils.file_organizer instead of importing from main
- Removed hacky sys.path manipulation

**Phase 11: Clean Up Unused Imports**
- Removed unused standard library imports (sqlite3, hashlib, shutil, etc.)
- Removed unused PyQt6 imports (QColor, QBrush, QThread, etc.)
- Removed unused third-party imports (xisf, re)
- Main file reduced from 809 to 291 lines (64% reduction)

**Impact:**
- Improved code maintainability and readability
- Eliminated code duplication
- Fixed circular dependency issues
- Cleaner import structure
- Better separation of concerns
