+++ AstroFileManager.py
@@ -5,25 +5,9 @@ A PyQt6-based application for cataloging, organizing, and managing XISF astropho
 """
 
 import sys
-import os
-import sqlite3
-import hashlib
-import shutil
-from pathlib import Path
-from datetime import datetime, timedelta
-from typing import Optional, Any
-from zoneinfo import ZoneInfo
-from PyQt6.QtWidgets import (
-    QApplication, QMainWindow, QWidget, QVBoxLayout, QHBoxLayout,
-    QPushButton, QFileDialog, QMessageBox, QTabWidget, QTableWidget,
-    QTableWidgetItem, QLabel, QProgressBar, QTextEdit, QTreeWidget,
-    QTreeWidgetItem, QRadioButton, QButtonGroup, QGroupBox, QComboBox,
-    QLineEdit, QMenu
-)
-from PyQt6.QtGui import QColor, QBrush
-from PyQt6.QtCore import Qt, QThread, pyqtSignal, QSettings
-import xisf
-import re
+from typing import Any
+from PyQt6.QtWidgets import QApplication, QMainWindow, QWidget, QVBoxLayout, QTabWidget
+from PyQt6.QtCore import Qt, QSettings
 
 # Import constants
 from constants import (
@@ -36,9 +20,6 @@ from constants import (
 from core.database import DatabaseManager
 from core.calibration import CalibrationMatcher
 
-# Import import/export modules
-from import_export.csv_exporter import CSVExporter
-
 # Import UI modules
 from ui.import_tab import ImportTab
 from ui.settings_tab import SettingsTab
@@ -46,77 +27,6 @@ from ui.maintenance_tab import MaintenanceTab
 from ui.sessions_tab import SessionsTab
 from ui.analytics_tab import AnalyticsTab
 from ui.view_catalog_tab import ViewCatalogTab
-
-
-def generate_organized_path(repo_path: str, obj: Optional[str], filt: Optional[str],
-                           imgtyp: Optional[str], exp: Optional[float], temp: Optional[float],
-                           xbin: Optional[int], ybin: Optional[int], date: Optional[str],
-                           original_filename: str) -> str:
-    """Generate the organized path and filename for a file"""
-    # Sanitize values
-    obj = obj or "Unknown"
-    filt = filt or "NoFilter"
-    imgtyp = imgtyp or "Unknown"
-    date = date or "0000-00-00"
-
-    # Determine binning string
-    if xbin and ybin:
-        try:
-            binning = f"Bin{int(float(xbin))}x{int(float(ybin))}"
-        except (ValueError, TypeError):
-            binning = "Bin1x1"
-    else:
-        binning = "Bin1x1"
-
-    # Determine temp string (round to nearest degree)
-    if temp is not None:
-        try:
-            temp_float = float(temp)
-            temp_str = f"{int(round(temp_float))}C"
-        except (ValueError, TypeError):
-            temp_str = "0C"
-    else:
-        temp_str = "0C"
-
-    # Extract sequence number from original filename if possible
-    seq_match = re.search(r'_(\d+)\.(xisf|fits?)$', original_filename, re.IGNORECASE)
-    seq = seq_match.group(1) if seq_match else "001"
-
-    # Determine file type and path structure
-    if 'light' in imgtyp.lower():
-        # Lights/[Object]/[Filter]/[filename]
-        subdir = os.path.join("Lights", obj, filt)
-        try:
-            exp_str = f"{int(float(exp))}s" if exp else "0s"
-        except (ValueError, TypeError):
-            exp_str = "0s"
-        new_filename = f"{date}_{obj}_{filt}_{exp_str}_{temp_str}_{binning}_{seq}.xisf"
-
-    elif 'dark' in imgtyp.lower():
-        # Calibration/Darks/[exp]_[temp]_[binning]/[filename]
-        try:
-            exp_str = f"{int(float(exp))}s" if exp else "0s"
-        except (ValueError, TypeError):
-            exp_str = "0s"
-        subdir = os.path.join("Calibration", "Darks", f"{exp_str}_{temp_str}_{binning}")
-        new_filename = f"{date}_Dark_{exp_str}_{temp_str}_{binning}_{seq}.xisf"
-
-    elif 'flat' in imgtyp.lower():
-        # Calibration/Flats/[date]/[filter]_[temp]_[binning]/[filename]
-        subdir = os.path.join("Calibration", "Flats", date, f"{filt}_{temp_str}_{binning}")
-        new_filename = f"{date}_Flat_{filt}_{temp_str}_{binning}_{seq}.xisf"
-
-    elif 'bias' in imgtyp.lower():
-        # Calibration/Bias/[temp]_[binning]/[filename]
-        subdir = os.path.join("Calibration", "Bias", f"{temp_str}_{binning}")
-        new_filename = f"{date}_Bias_{temp_str}_{binning}_{seq}.xisf"
-
-    else:
-        # Unknown type - put in root with original structure
-        subdir = "Uncategorized"
-        new_filename = original_filename
-
-    return os.path.join(repo_path, subdir, new_filename)
 
 
 class XISFCatalogGUI(QMainWindow):
@@ -184,435 +94,7 @@ class XISFCatalogGUI(QMainWindow):
         
         # Connect tab change to refresh
         tabs.currentChanged.connect(self.on_tab_changed)
-    
 
-    def on_keyword_changed(self) -> None:
-        """Update the current value dropdown when keyword selection changes"""
-        keyword = self.keyword_combo.currentText()
-        self.populate_current_values(keyword)
-
-    def populate_current_values(self, keyword: str) -> None:
-        """Populate the current value dropdown with existing values from the database"""
-        try:
-            conn = sqlite3.connect(self.db_path)
-            cursor = conn.cursor()
-            
-            # Map FITS keywords to database column names
-            column_map = {
-                'TELESCOP': 'telescop',
-                'INSTRUME': 'instrume',
-                'OBJECT': 'object',
-                'FILTER': 'filter',
-                'IMAGETYP': 'imagetyp',
-                'DATE-LOC': 'date_loc'
-            }
-            
-            column = column_map.get(keyword)
-            if column:
-                cursor.execute(f'SELECT DISTINCT {column} FROM xisf_files WHERE {column} IS NOT NULL ORDER BY {column}')
-                values = [row[0] for row in cursor.fetchall()]
-                
-                self.current_value_combo.clear()
-                self.current_value_combo.addItems(values)
-            
-            conn.close()
-        except Exception as e:
-            QMessageBox.critical(self, 'Error', f'Failed to load values: {e}')
-    
-    def replace_values(self) -> None:
-        """Replace values in the database"""
-        keyword = self.keyword_combo.currentText()
-        current_value = self.current_value_combo.currentText()
-        replacement_value = self.replacement_input.text()
-        
-        if not current_value:
-            QMessageBox.warning(self, 'No Value Selected', 'Please select a current value to replace.')
-            return
-        
-        if not replacement_value:
-            QMessageBox.warning(self, 'No Replacement', 'Please enter a replacement value.')
-            return
-        
-        # Confirm the replacement
-        reply = QMessageBox.question(
-            self, 'Confirm Replacement',
-            f'Replace all occurrences of:\n\n'
-            f'Keyword: {keyword}\n'
-            f'Current Value: "{current_value}"\n'
-            f'Replacement Value: "{replacement_value}"\n\n'
-            f'Are you sure?',
-            QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No,
-            QMessageBox.StandardButton.No
-        )
-        
-        if reply == QMessageBox.StandardButton.Yes:
-            try:
-                conn = sqlite3.connect(self.db_path)
-                cursor = conn.cursor()
-                
-                # Map FITS keywords to database column names
-                column_map = {
-                    'TELESCOP': 'telescop',
-                    'INSTRUME': 'instrume',
-                    'OBJECT': 'object',
-                    'FILTER': 'filter',
-                    'IMAGETYP': 'imagetyp',
-                    'DATE-LOC': 'date_loc'
-                }
-                
-                column = column_map.get(keyword)
-                if column:
-                    cursor.execute(f'UPDATE xisf_files SET {column} = ? WHERE {column} = ?', 
-                                   (replacement_value, current_value))
-                    rows_affected = cursor.rowcount
-                    conn.commit()
-                    
-                    QMessageBox.information(
-                        self, 'Success',
-                        f'Successfully replaced {rows_affected} occurrence(s).'
-                    )
-                    
-                    # Refresh the current values dropdown
-                    self.populate_current_values(keyword)
-                    self.replacement_input.clear()
-                
-                conn.close()
-            except Exception as e:
-                QMessageBox.critical(self, 'Error', f'Failed to replace values: {e}')
-
-    def preview_organization(self) -> None:
-        """Preview the file organization plan"""
-        repo_path = self.settings.value('repository_path', '')
-        
-        if not repo_path:
-            QMessageBox.warning(
-                self, 'No Repository Path',
-                'Please set the repository path in the Settings tab first.'
-            )
-            return
-        
-        self.organize_log.clear()
-        self.organize_log.append("Generating organization preview...\n")
-        
-        try:
-            conn = sqlite3.connect(self.db_path)
-            cursor = conn.cursor()
-            
-            # Get all files
-            cursor.execute('''
-                SELECT filepath, filename, object, filter, imagetyp, 
-                       exposure, ccd_temp, xbinning, ybinning, date_loc
-                FROM xisf_files
-                ORDER BY object, filter, date_loc
-            ''')
-            
-            files = cursor.fetchall()
-            conn.close()
-            
-            if not files:
-                self.organize_log.append("No files found in database.")
-                return
-            
-            self.organize_log.append(f"Found {len(files)} files to organize.\n")
-            self.organize_log.append("Sample organization plan (showing first 10):\n")
-            
-            for i, (filepath, filename, obj, filt, imgtyp, exp, temp, xbin, ybin, date) in enumerate(files[:10]):
-                new_path = generate_organized_path(
-                    repo_path, obj, filt, imgtyp, exp, temp, xbin, ybin, date, filename
-                )
-                self.organize_log.append(f"\nFrom: {filepath}")
-                self.organize_log.append(f"To:   {new_path}")
-            
-            if len(files) > 10:
-                self.organize_log.append(f"\n... and {len(files) - 10} more files")
-            
-            self.organize_log.append("\n" + "="*60)
-            self.organize_log.append("This is a preview only. No files have been moved.")
-            
-        except Exception as e:
-            self.organize_log.append(f"\nError generating preview: {e}")
-    
-    def execute_organization(self) -> None:
-        """Execute the file organization"""
-        repo_path = self.settings.value('repository_path', '')
-        
-        if not repo_path:
-            QMessageBox.warning(
-                self, 'No Repository Path',
-                'Please set the repository path in the Settings tab first.'
-            )
-            return
-        
-        reply = QMessageBox.question(
-            self, 'Confirm File Organization',
-            'This will copy all files to the new organized structure and update the database.\n\n'
-            'Original files will NOT be deleted.\n\n'
-            'This may take some time. Continue?',
-            QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No,
-            QMessageBox.StandardButton.No
-        )
-        
-        if reply != QMessageBox.StandardButton.Yes:
-            return
-        
-        self.organize_log.clear()
-        self.organize_log.append("Starting file organization...\n")
-        
-        try:
-            import shutil
-            
-            conn = sqlite3.connect(self.db_path)
-            cursor = conn.cursor()
-            
-            # Get all files
-            cursor.execute('''
-                SELECT id, filepath, filename, object, filter, imagetyp, 
-                       exposure, ccd_temp, xbinning, ybinning, date_loc
-                FROM xisf_files
-                ORDER BY object, filter, date_loc
-            ''')
-            
-            files = cursor.fetchall()
-            
-            if not files:
-                self.organize_log.append("No files found in database.")
-                conn.close()
-                return
-            
-            success_count = 0
-            error_count = 0
-            
-            for file_id, filepath, filename, obj, filt, imgtyp, exp, temp, xbin, ybin, date in files:
-                try:
-                    # Check if source file exists
-                    if not os.path.exists(filepath):
-                        self.organize_log.append(f"❌ Source not found: {filepath}")
-                        error_count += 1
-                        continue
-                    
-                    # Generate new path and filename
-                    new_path = generate_organized_path(
-                        repo_path, obj, filt, imgtyp, exp, temp, xbin, ybin, date, filename
-                    )
-                    
-                    # Create directory if it doesn't exist
-                    os.makedirs(os.path.dirname(new_path), exist_ok=True)
-                    
-                    # Copy file if it doesn't already exist at destination
-                    if os.path.exists(new_path):
-                        self.organize_log.append(f"⚠️  Already exists: {new_path}")
-                    else:
-                        shutil.copy2(filepath, new_path)
-                        self.organize_log.append(f"✓ Copied: {os.path.basename(new_path)}")
-                    
-                    # Update database with new path and filename
-                    new_filename = os.path.basename(new_path)
-                    cursor.execute('UPDATE xisf_files SET filepath = ?, filename = ? WHERE id = ?',
-                                   (new_path, new_filename, file_id))
-                    success_count += 1
-                    
-                except Exception as e:
-                    self.organize_log.append(f"❌ Error with {filename}: {e}")
-                    error_count += 1
-            
-            conn.commit()
-            conn.close()
-            
-            self.organize_log.append("\n" + "="*60)
-            self.organize_log.append(f"Organization complete!")
-            self.organize_log.append(f"Successfully organized: {success_count}")
-            self.organize_log.append(f"Errors: {error_count}")
-            
-            QMessageBox.information(
-                self, 'Organization Complete',
-                f'Successfully organized {success_count} files.\n'
-                f'Errors: {error_count}\n\n'
-                'Check the log for details.'
-            )
-            
-        except Exception as e:
-            self.organize_log.append(f"\nFatal error: {e}")
-            QMessageBox.critical(self, 'Error', f'Failed to organize files: {e}')
-
-    def create_settings_tab(self) -> QWidget:
-        """Create the settings tab"""
-        widget = QWidget()
-        layout = QVBoxLayout(widget)
-        
-        # Image Repository Location
-        repo_group = QGroupBox("Image Repository")
-        repo_layout = QVBoxLayout()
-        
-        repo_info = QLabel("Set the location for your organized XISF files:")
-        repo_layout.addWidget(repo_info)
-        
-        repo_path_layout = QHBoxLayout()
-        repo_path_label = QLabel("Repository Path:")
-        repo_path_label.setMinimumWidth(120)
-        self.repo_path_input = QLineEdit()
-        self.repo_path_input.setReadOnly(True)
-        current_repo = self.settings.value('repository_path', '')
-        if current_repo:
-            # Standardize on forward slashes
-            current_repo = current_repo.replace('\\', '/')
-        self.repo_path_input.setText(current_repo)
-        
-        browse_repo_btn = QPushButton('Browse...')
-        browse_repo_btn.clicked.connect(self.browse_repository)
-        
-        repo_path_layout.addWidget(repo_path_label)
-        repo_path_layout.addWidget(self.repo_path_input)
-        repo_path_layout.addWidget(browse_repo_btn)
-        repo_layout.addLayout(repo_path_layout)
-        
-        repo_group.setLayout(repo_layout)
-        layout.addWidget(repo_group)
-
-        # Timezone settings group
-        timezone_group = QGroupBox("Timezone")
-        timezone_layout = QVBoxLayout()
-
-        timezone_info = QLabel("Set your local timezone for DATE-OBS conversion:")
-        timezone_layout.addWidget(timezone_info)
-
-        timezone_help = QLabel("Used to convert UTC timestamps (DATE-OBS) to local time for session grouping.")
-        timezone_help.setStyleSheet("color: #888888; font-size: 10px;")
-        timezone_layout.addWidget(timezone_help)
-
-        timezone_selector_layout = QHBoxLayout()
-        timezone_label = QLabel("Timezone:")
-        timezone_label.setMinimumWidth(120)
-        self.timezone_combo = QComboBox()
-
-        # Common timezones
-        common_timezones = [
-            'UTC',
-            'America/New_York',
-            'America/Chicago',
-            'America/Denver',
-            'America/Los_Angeles',
-            'America/Phoenix',
-            'America/Anchorage',
-            'Pacific/Honolulu',
-            'Europe/London',
-            'Europe/Paris',
-            'Europe/Berlin',
-            'Europe/Rome',
-            'Europe/Madrid',
-            'Europe/Athens',
-            'Asia/Tokyo',
-            'Asia/Shanghai',
-            'Asia/Hong_Kong',
-            'Asia/Singapore',
-            'Asia/Dubai',
-            'Australia/Sydney',
-            'Australia/Melbourne',
-            'Australia/Perth',
-            'Pacific/Auckland'
-        ]
-
-        self.timezone_combo.addItems(common_timezones)
-
-        # Set current timezone
-        current_timezone = self.settings.value('timezone', 'UTC')
-        index = self.timezone_combo.findText(current_timezone)
-        if index >= 0:
-            self.timezone_combo.setCurrentIndex(index)
-
-        save_timezone_btn = QPushButton('Save Timezone')
-        save_timezone_btn.clicked.connect(self.save_timezone_setting)
-
-        timezone_selector_layout.addWidget(timezone_label)
-        timezone_selector_layout.addWidget(self.timezone_combo)
-        timezone_selector_layout.addWidget(save_timezone_btn)
-        timezone_layout.addLayout(timezone_selector_layout)
-
-        timezone_group.setLayout(timezone_layout)
-        layout.addWidget(timezone_group)
-
-        # Theme settings group
-        theme_group = QGroupBox("Theme")
-        theme_layout = QVBoxLayout()
-        
-        # Radio buttons for theme selection
-        self.theme_button_group = QButtonGroup()
-        self.standard_theme_radio = QRadioButton("Standard Theme")
-        self.dark_theme_radio = QRadioButton("Dark Theme")
-        
-        self.theme_button_group.addButton(self.standard_theme_radio, 0)
-        self.theme_button_group.addButton(self.dark_theme_radio, 1)
-        
-        theme_layout.addWidget(self.standard_theme_radio)
-        theme_layout.addWidget(self.dark_theme_radio)
-        
-        # Set current theme
-        current_theme = self.settings.value('theme', 'dark')
-        if current_theme == 'standard':
-            self.standard_theme_radio.setChecked(True)
-        else:
-            self.dark_theme_radio.setChecked(True)
-        
-        theme_group.setLayout(theme_layout)
-        layout.addWidget(theme_group)
-        
-        # OK button
-        button_layout = QHBoxLayout()
-        button_layout.addStretch()
-        ok_button = QPushButton('Apply Theme')
-        ok_button.clicked.connect(self.apply_theme_setting)
-        button_layout.addWidget(ok_button)
-        layout.addLayout(button_layout)
-        
-        # Add stretch to push everything to the top
-        layout.addStretch()
-        
-        return widget
-    
-    def browse_repository(self) -> None:
-        """Browse for repository location"""
-        current_path = self.repo_path_input.text()
-        directory = QFileDialog.getExistingDirectory(
-            self, 'Select Image Repository Location', 
-            current_path if current_path else ''
-        )
-        
-        if directory:
-            # Normalize path to use OS-specific separators
-            directory = os.path.normpath(directory)
-            self.repo_path_input.setText(directory)
-            self.settings.setValue('repository_path', directory)
-            QMessageBox.information(
-                self, 'Repository Path Updated',
-                f'Image repository location set to:\n{directory}'
-            )
-    
-    def save_timezone_setting(self) -> None:
-        """Save the selected timezone"""
-        timezone = self.timezone_combo.currentText()
-        self.settings.setValue('timezone', timezone)
-        QMessageBox.information(
-            self,
-            'Timezone Saved',
-            f'Timezone set to: {timezone}\n\nThis will be used for converting DATE-OBS timestamps.'
-        )
-
-    def apply_theme_setting(self) -> None:
-        """Apply the selected theme"""
-        if self.standard_theme_radio.isChecked():
-            theme = 'standard'
-        else:
-            theme = 'dark'
-
-        # Save theme preference
-        self.settings.setValue('theme', theme)
-
-        # Show message that restart is needed
-        QMessageBox.information(
-            self,
-            'Theme Changed',
-            'Theme has been changed. Please restart the application for the changes to take effect.'
-        )
-    
     def connect_signals(self) -> None:
         """Connect signals after all widgets are created"""
         # Connect column resize signals to save settings
